Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -5,7 +5,7 @@
 // }
 
 int main() {
-    int number;
+    float number;
     scanf("%f",&number);
     printf("You entered: %f",number);
     return 0;
